dashboard/requirements-dev.txt
factory-boy==3.2.0
ipdb==0.13.5
mypy==0.812
prospector==1.3.1
pytest==6.2.2
pytest-django==4.1.0
